requirements.txt
absl-py==2.0.0
aiohttp==3.8.1
aiosignal==1.3.1
appdirs==1.4.4
asciitree==0.3.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==2.0.0
configmypy==0.1.0
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
cytoolz==0.12.0
dask==2022.7.0
Deprecated==1.2.14
docker-pycreds==0.4.0
entrypoints==0.4
exceptiongroup==1.1.1
fasteners==0.18
filelock==3.9.0
flit_core==3.6.0
fonttools==4.25.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.3
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
gmpy2==2.1.2
google-auth==2.23.2
google-auth-oauthlib==1.0.0
grpcio==1.48.2
h5py==3.7.0
idna==3.4
imagecodecs==2021.8.26
imageio==2.26.0
importlib-metadata==6.8.0
importlib-resources==5.2.0
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.1.1
kiwisolver==1.4.4
locket==1.0.0
Markdown==3.4.4
MarkupSafe==2.1.1
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.2
munkres==1.1.4
networkx==2.8.4
neuraloperator==0.1.3
nose==1.3.7
numcodecs==0.11.0
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
openreview-py==1.32.3
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
partd==1.2.0
pathtools==0.1.2
Pillow==9.4.0
pip==23.0.1
plotly==5.22.0
pluggy==1.0.0
ply==3.11
pooch==1.4.0
protobuf==4.22.3
psutil==5.9.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pyhocon==0.3.60
PyJWT==2.8.0
pylatexenc==2.10
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks==1.7.1
pytest==7.3.0
pytest-mock==3.10.0
python-dateutil==2.8.2
pytz==2022.7
pyu2f==0.1.5
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
sentry-sdk==1.19.1
setproctitle==1.3.2
setuptools==69.5.1
sip==6.6.2
six==1.16.0
smmap==5.0.0
soupsieve==2.5
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.14.1
tensorboard-data-server==0.7.0
tensorly==0.8.0
tensorly-torch==0.4.0
threadpoolctl==2.2.0
tifffile==2021.7.2
tld==0.13
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.0
torchdiffeq==0.2.3
torchvision==0.15.0
tornado==6.2
tqdm==4.64.1
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.14
wandb==0.14.2
Werkzeug==3.0.0
wheel==0.38.4
wrapt==1.16.0
yacs==0.1.8
yarl==1.7.2
zarr==2.14.2
zipp==3.11.0
